Refactored structure of for loop due to ansible test failures in python 2.6
